完成權限控制相關的所有頁面調整.

+++ src/routes/03RightsManagement/personManage/index.js
@@ -4,6 +4,7 @@ import moment from "moment";
 import { Table, Row, Col, Card, Form, Input, Select, Button, Switch, Divider, Modal, Icon, message } from "antd";
 import styles from "../../TableList.less";
 import config from "../../../utils/config";
+import utils from "../../../utils";
 
 const { Option } = Select;
 
@@ -90,13 +91,23 @@ const PersonManage = ({
             dataIndex: "",
             width: 150,
             render: (text, record) => {
-                return (
-                    <Fragment>
-                        <a onClick={editModalShow.bind(this, record)}>编辑</a>
-                        <Divider type="vertical" />
-                        <a onClick={deleteItem.bind(this, record)}>刪除</a>
-                    </Fragment>
-                );
+                if (utils.inPages("24")) {
+                    return (
+                        <Fragment>
+                            <a onClick={editModalShow.bind(this, record)}>编辑</a>
+                            <Divider type="vertical" />
+                            <a onClick={deleteItem.bind(this, record)}>刪除</a>
+                        </Fragment>
+                    );
+                } else {
+                    return (
+                        <Fragment>
+                            <a href={"javascript:return false;"} style={{ opacity: 0.2, color: "#272727" }}>编辑</a>
+                            <Divider type="vertical" />
+                            <a href={"javascript:return false;"} style={{ opacity: 0.2, color: "#272727" }}>刪除</a>
+                        </Fragment>
+                    );
+                }
             }
         }
     ];
